requirements.txt
certifi==2022.12.7
jmespath==1.0.1
polygon-api-client==1.10.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
stockstats==0.6.0
tzdata==2023.3
urllib3==1.26.16
websockets==11.0.3